pyproject.toml
[tool.poetry]
name = "log-generator"
version = "0.1.0"
description = "A log generator in Python that uses the Flask framework."
authors = ["joaovictorvns <jv369t7d2hx@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.0.3"
gunicorn = "^23.0.0"
dynaconf = "^3.2.6"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
pytest = "^8.3.3"
pytest-flask = "^1.3.0"
pytest-cov = "^5.0.0"
selenium = "^4.27.1"
webdriver-manager = "^4.0.2"
mypy = "^1.11.2"
flake8 = "^7.1.1"
pylint = "^3.3.1"
isort = "^5.13.2"
ruff = "^0.6.9"
taskipy = "^1.13.0"

[tool.poetry.group.doc.dependencies]
mkdocs = "^1.6.1"
mkdocstrings = "^0.26.1"
mkdocstrings-python = "^1.11.1"
mkdocs-material = "^9.5.40"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.ruff]
line-length = 79
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "tests",
]

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "C90",
    "I",
    "N",
    "D",
    "UP",
    "PL",
    "RUF",
]
preview = true
ignore = []

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "single"

[tool.taskipy.tasks]
lint = "ruff check . && flake8 . && pylint ."
format = "ruff format --check --diff . && isort --check --diff ."
pre_test = "mypy . && task lint && task format"
test = "pytest --cov=log_generator -vv"
post_test = "coverage html"
